+++ flowman-core/src/main/scala/com/dimajix/flowman/transforms/ProjectTransformer.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Kaya Kupferschmidt
+ * Copyright 2019-2022 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -32,7 +32,7 @@ object ProjectTransformer {
         dtype:Option[FieldType]=None
     )
 }
-case class ProjectTransformer(
+final case class ProjectTransformer(
     columns:Seq[ProjectTransformer.Column]
 ) extends Transformer {
     /**