+++ flowman-core/src/main/scala/com/dimajix/flowman/transforms/TypeReplacer.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Kaya Kupferschmidt
+ * Copyright 2019-2022 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -38,7 +38,7 @@ import com.dimajix.flowman.types.VarcharType
   * This
   * @param replace
   */
-case class TypeReplacer(replace:Map[String, FieldType]) extends Transformer {
+final case class TypeReplacer(replace:Map[String, FieldType]) extends Transformer {
     private val typeAliases = Map(
         "text" -> "string",
         "long" -> "bigint",